requirements.txt
sentence-transformers==3.3.1
python-dotenv==1.0.1
scikit-learn==1.3.0
requests
python-dotenv
pdfminer.six
pandas
nump
openai
spacy
fuzzywuzzy
language-tool-python

